pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["Sara Zwairi <sara.zwairi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
djangorestframework-simplejwt = "^4.8.0"
whitenoise = "^5.3.0"
gunicorn = "^20.1.0"
django-cors-headers = "^3.8.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
pytest = "^6.2.5"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
